Fixes infinite loop bug
* If list had one node and was a cycle, crash would occur
test suit would not run. I believe because program was stuck
in a loop.

* I tried to use p2 pointer as my stopping condition because
it would find null just a little quicker than its sister node, p1.
p1 was moving at half the pace. For a large data set, this matters.
I have to make sure I am not skipping over the null. This was a
a suspision of mine when writing the program. I should have included
the if statment, and removed later if needed.